# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Fasal Agriculture Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SensorDataInput {
  nitrogen: number;
  phosphorus: number;
  potassium: number;
  ph: number;
  moisture: number;
  /** @nullable */
  deviceId?: string | null;
}

export interface SensorData {
  id: number;
  nitrogen: number;
  phosphorus: number;
  potassium: number;
  ph: number;
  moisture: number;
  /** @nullable */
  deviceId?: string | null;
  createdAt: string;
}

export interface AiRecommendationInput {
  nitrogen: number;
  phosphorus: number;
  potassium: number;
  ph: number;
  moisture: number;
  /** @nullable */
  temperature?: number | null;
  /** @nullable */
  humidity?: number | null;
  /** @nullable */
  rainfall?: number | null;
}

export interface AiRecommendation {
  id: number;
  fertilizerAdvice: string;
  irrigationSuggestion: string;
  riskAnalysis: string;
  cropHealthPercent: number;
  riskLevel: string;
  yieldPercent: number;
  nitrogen: number;
  phosphorus: number;
  potassium: number;
  ph: number;
  moisture: number;
  createdAt: string;
}

export interface DiseaseDetectInput {
  imageDescription: string;
  /** @nullable */
  cropName?: string | null;
}

export interface DiseaseDetectResult {
  plantName: string;
  diseaseName: string;
  confidencePercent: number;
  treatment: string;
  severity: string;
}

export interface WeatherData {
  temperature: number;
  humidity: number;
  rainfall: number;
  description: string;
  location: string;
  windSpeed: number;
  feelsLike: number;
  updatedAt: string;
}

export interface InsuranceRisk {
  riskLevel: string;
  riskScore: number;
  eligibleForClaim: boolean;
  reasons: string[];
  recommendations: string[];
}

export interface InsuranceClaim {
  id: number;
  claimType: string;
  riskLevel: string;
  description: string;
  status: string;
  rewardPoints: number;
  /** @nullable */
  walletAddress?: string | null;
  createdAt: string;
}

export interface CreateInsuranceClaimInput {
  claimType: string;
  description: string;
  /** @nullable */
  walletAddress?: string | null;
}

export interface MarketPrice {
  id: number;
  crop: string;
  price: number;
  unit: string;
  market: string;
  state: string;
  change: number;
  updatedAt: string;
}

export interface MarketListing {
  id: number;
  title: string;
  description: string;
  crop: string;
  price: number;
  quantity: number;
  unit: string;
  sellerName: string;
  /** @nullable */
  sellerWallet?: string | null;
  location: string;
  status: string;
  createdAt: string;
}

export interface CreateMarketListingInput {
  title: string;
  description: string;
  crop: string;
  price: number;
  quantity: number;
  unit: string;
  sellerName: string;
  /** @nullable */
  sellerWallet?: string | null;
  location: string;
}

export interface ProductRecommendation {
  id: number;
  name: string;
  category: string;
  description: string;
  price: number;
  reason: string;
  rating: number;
}

export interface Comment {
  id: number;
  author: string;
  content: string;
  createdAt: string;
}

export interface CommunityPost {
  id: number;
  author: string;
  /** @nullable */
  walletAddress?: string | null;
  content: string;
  visibility: string;
  likes: number;
  /** @nullable */
  filecoinCid?: string | null;
  /** @nullable */
  badge?: string | null;
  comments: Comment[];
  createdAt: string;
}

export interface CreateCommunityPostInput {
  author: string;
  /** @nullable */
  walletAddress?: string | null;
  content: string;
  visibility?: string;
  /** @nullable */
  badge?: string | null;
}

export interface AddCommentInput {
  author: string;
  content: string;
}

export interface ChatMessage {
  id: number;
  sender: string;
  content: string;
  encryptedContent: string;
  isEncrypted: boolean;
  createdAt: string;
}

export interface SendMessageInput {
  sender: string;
  content: string;
}

export interface Expert {
  id: number;
  name: string;
  specialization: string;
  experience: string;
  rating: number;
  questionsAnswered: number;
  badge: string;
}

export interface ExpertQuestion {
  id: number;
  question: string;
  askedBy: string;
  /** @nullable */
  expertId?: number | null;
  /** @nullable */
  expertName?: string | null;
  /** @nullable */
  answer?: string | null;
  status: string;
  createdAt: string;
}

export interface AskExpertInput {
  question: string;
  askedBy: string;
  /** @nullable */
  expertId?: number | null;
}

export type FilecoinStoreInputData = { [key: string]: unknown };

export interface FilecoinStoreInput {
  dataType: string;
  data: FilecoinStoreInputData;
}

export interface FilecoinResult {
  cid: string;
  url: string;
  dataType: string;
  storedAt: string;
}

export interface FilecoinRecord {
  id: number;
  cid: string;
  url: string;
  dataType: string;
  createdAt: string;
}

export interface RewardTransaction {
  id: number;
  activity: string;
  points: number;
  createdAt: string;
}

export interface RewardsInfo {
  /** @nullable */
  walletAddress?: string | null;
  totalPoints: number;
  badges: string[];
  transactions: RewardTransaction[];
}

export interface ConnectWalletInput {
  walletAddress: string;
}

export interface AddRewardInput {
  activity: string;
  points: number;
  /** @nullable */
  walletAddress?: string | null;
}

export interface EventLog {
  id: number;
  eventType: string;
  description: string;
  /** @nullable */
  metadata?: string | null;
  createdAt: string;
}

export interface AnalyticsSummary {
  totalSensorReadings: number;
  totalAiRecommendations: number;
  totalInsuranceClaims: number;
  totalCommunityPosts: number;
  totalMarketListings: number;
  avgCropHealth: number;
  avgSoilPh: number;
  avgMoisture: number;
  recentLogs: EventLog[];
}

export type GetSensorHistoryParams = {
  limit?: number;
};

export type GetWeatherParams = {
  lat?: number;
  lon?: number;
};