# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Fasal Agriculture Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai-recommendation:
    post:
      operationId: getAiRecommendation
      tags: [ai]
      summary: Get AI-powered crop recommendation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiRecommendationInput"
      responses:
        "200":
          description: AI recommendation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiRecommendation"
...
  /disease-detect:
    post:
      operationId: detectDisease
      tags: [ai]
      summary: Detect crop disease from image description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiseaseDetectInput"
      responses:
        "200":
          description: Disease detection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiseaseDetectResult"
...
  /community/posts:
    get:
      operationId: getCommunityPosts
      tags: [community]
      summary: Get community feed posts
      responses:
        "200":
          description: Community posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommunityPost"
    post:
      operationId: createCommunityPost
      tags: [community]
      summary: Create a community post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommunityPostInput"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPost"
  /community/posts/{id}/like:
    post:
      operationId: likeCommunityPost
      tags: [community]
      summary: Like a community post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPost"
  /community/posts/{id}/comments:
    post:
      operationId: addCommentToPost
      tags: [community]
      summary: Add comment to post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCommentInput"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPost"
  /community/messages:
    get:
      operationId: getCommunityMessages
      tags: [community]
      summary: Get community chat messages
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendCommunityMessage
      tags: [community]
      summary: Send an encrypted community message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
...
  /community/expert-questions:
    get:
      operationId: getExpertQuestions
      tags: [community]
      summary: Get questions asked to experts
      responses:
        "200":
          description: Expert questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpertQuestion"
    post:
      operationId: askExpertQuestion
      tags: [community]
      summary: Ask an expert a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskExpertInput"
      responses:
        "201":
          description: Question submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpertQuestion"
...
components:
  schemas:
...
    AiRecommendationInput:
      type: object
      properties:
        nitrogen:
          type: number
        phosphorus:
          type: number
        potassium:
          type: number
        ph:
          type: number
        moisture:
          type: number
        temperature:
          type: ["number", "null"]
        humidity:
          type: ["number", "null"]
        rainfall:
          type: ["number", "null"]
      required:
        - nitrogen
        - phosphorus
        - potassium
        - ph
        - moisture
    AiRecommendation:
      type: object
      properties:
        id:
          type: integer
        fertilizerAdvice:
          type: string
        irrigationSuggestion:
          type: string
        riskAnalysis:
          type: string
        cropHealthPercent:
          type: number
        riskLevel:
          type: string
        yieldPercent:
          type: number
        nitrogen:
          type: number
        phosphorus:
          type: number
        potassium:
          type: number
        ph:
          type: number
        moisture:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fertilizerAdvice
        - irrigationSuggestion
        - riskAnalysis
        - cropHealthPercent
        - riskLevel
        - yieldPercent
        - nitrogen
        - phosphorus
        - potassium
        - ph
        - moisture
        - createdAt
    DiseaseDetectInput:
      type: object
      properties:
        imageDescription:
          type: string
        cropName:
          type: ["string", "null"]
      required:
        - imageDescription
    DiseaseDetectResult:
      type: object
      properties:
        plantName:
          type: string
        diseaseName:
          type: string
        confidencePercent:
          type: number
        treatment:
          type: string
        severity:
          type: string
      required:
        - plantName
        - diseaseName
        - confidencePercent
        - treatment
        - severity
...
    CommunityPost:
      type: object
      properties:
        id:
          type: integer
        author:
          type: string
        walletAddress:
          type: ["string", "null"]
        content:
          type: string
        visibility:
          type: string
        likes:
          type: integer
        filecoinCid:
          type: ["string", "null"]
        badge:
          type: ["string", "null"]
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - author
        - content
        - visibility
        - likes
        - comments
        - createdAt
    Comment:
      type: object
      properties:
        id:
          type: integer
        author:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - author
        - content
        - createdAt
    CreateCommunityPostInput:
      type: object
      properties:
        author:
          type: string
        walletAddress:
          type: ["string", "null"]
        content:
          type: string
        visibility:
          type: string
          default: public
        badge:
          type: ["string", "null"]
      required:
        - author
        - content
    AddCommentInput:
      type: object
      properties:
        author:
          type: string
        content:
          type: string
      required:
        - author
        - content
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        sender:
          type: string
        content:
          type: string
        encryptedContent:
          type: string
        isEncrypted:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sender
        - content
        - encryptedContent
        - isEncrypted
        - createdAt
    SendMessageInput:
      type: object
      properties:
        sender:
          type: string
        content:
          type: string
      required:
        - sender
        - content
...
    ExpertQuestion:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        askedBy:
          type: string
        expertId:
          type: ["integer", "null"]
        expertName:
          type: ["string", "null"]
        answer:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - question
        - askedBy
        - status
        - createdAt
    AskExpertInput:
      type: object
      properties:
        question:
          type: string
        askedBy:
          type: string
        expertId:
          type: ["integer", "null"]
      required:
        - question
        - askedBy
...