# lib/api-zod/src/generated/types/aiRecommendationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Fasal Agriculture Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AiRecommendationInput {
  nitrogen: number;
  phosphorus: number;
  potassium: number;
  ph: number;
  moisture: number;
  /** @nullable */
  temperature?: number | null;
  /** @nullable */
  humidity?: number | null;
  /** @nullable */
  rainfall?: number | null;
}